+++ functions/soft.py
@@ -3,10 +3,10 @@ from chainer import backend
 from chainer.functions.evaluation.accuracy import Accuracy
 import numpy as np
 
-def hard_to_soft(label, classes=1000):
+def hard_to_soft(label, classes=1000, dtype=np.float32):
     result = [0.1 / (classes-1)] * classes
     result[label] = 0.9
-    return np.array(result).astype(np.float32)
+    return np.array(result).astype(dtype)
 
 def soft_softmax_cross_entropy(preds, soft_labels):
     """
